Final changes ID filtration

+++ app/Http/Controllers/TasksController.php
@@ -10,6 +10,23 @@ use App\Tasks;
 
 class TasksController extends Controller
 {
+    public function isThisTasksYours($id)
+    {
+        $zadanie = Tasks::findOrFail($id);
+
+        $live_user = Auth::id();
+        $task_user = $zadanie->user_id;
+
+        if($live_user == $task_user)
+        {
+            return true;
+        }
+        else
+        {
+            return false;
+        }
+    }
+
     public function logout()
     {
         Auth::logout();
@@ -21,7 +38,7 @@ class TasksController extends Controller
         //$zadania = \App\Tasks::all();
         //return view('tasks', ['zadania'=> $zadania]);
 
-        $zadania = Tasks::all();
+        $zadania = Tasks::where('user_id', Auth::id())->get();
         return view('tasks', compact('zadania'));
     }
 
@@ -34,13 +51,16 @@ class TasksController extends Controller
 
     public function edit($id)
     {
-        $zadanie = Tasks::findOrFail($id);
+        /*$zadanie = Tasks::findOrFail($id);
 
         $live_user = Auth::id();
         $task_user = $zadanie->user_id;
 
         if($live_user==$task_user)
+        */
+        if($this->isThisTasksYours($id))
         {
+            $zadanie = Tasks::findOrFail($id);
             return view('edit', compact('zadanie'));
         }
         else // if user will try to edit not his task...
@@ -62,6 +82,10 @@ class TasksController extends Controller
 
     public function done($id)
     {
+        if(!$this->isThisTasksYours($id))
+        {
+            return redirect ('/tasks');
+        }
         $zadanie = Tasks::findOrFail($id);
         $zadanie->status = true;
         $zadanie->save();
@@ -70,6 +94,11 @@ class TasksController extends Controller
 
     public function undone($id)
     {
+        if(!$this->isThisTasksYours($id))
+        {
+            return redirect ('/tasks');
+        }
+
         $zadanie = Tasks::findOrFail($id);
         $zadanie->status = false;
         $zadanie->save();
@@ -78,6 +107,11 @@ class TasksController extends Controller
 
     public function destroy($id)
     {
+        if(!$this->isThisTasksYours($id))
+        {
+            return redirect ('/tasks');
+        }
+
         Tasks::findOrFail($id)->delete();
         return redirect('/tasks');
     }
